requirements.txt
discord.py==2.4.0
aiohttp
python-dotenv
openai>=1.30.0
Pillow
numpy
requests
aiosqlite
asyncio
loguru